+++ packages/mangrove.js/src/types/index.ts
@@ -32,6 +32,7 @@ export interface MarketParams {
 
 export type Bigish = Big | number | string;
 
-export type TradeParams =
+export type TradeParams = { slippage?: number } & (
   | { volume: Bigish; price: Bigish }
-  | { wants: Bigish; gives: Bigish };
+  | { wants: Bigish; gives: Bigish }
+);